test(collector): isolate body-limit route tests
