+++ .agents/skills/verify-dns-ops/features/collector.request-body-limits.md
@@ -9,7 +9,7 @@ paths:
   - apps/collector/src/jobs/fleet-report.ts
   - apps/collector/src/jobs/fleet-report.test.ts
   - apps/collector/src/jobs/probe-routes.ts
-  - apps/collector/src/jobs/probe-routes.test.ts
+  - apps/collector/src/middleware/request-body-limit.routes.test.ts
 always_with: []
 ---
 # Collector request-body limits
