--- verification/receipts/20260902-163914Z-5c68f84-fresh--collector.request-body-limits.md
@@ -1,74 +1,0 @@
----
-receipt: verification-receipt/v0
-run_id: 20260902-163914Z-5c68f84-fresh
-feature_id: collector.request-body-limits
-profile: critical
-surface: api
-sha: 5c68f8446aaf71ae47c82e60c81347af059f1169
-code_digest: 40bda71eee2abe3f01795fb2552feefee6b2a2c9455b218ca490d3b923402c17
-dirty: false
-untracked: 0
-status: passed
-reason: ""
-verifier: fresh
-verifier_session: "pi-fresh:01a062fc-c548-7aed-92a7-a621121b9826"
-evidence_dir: verification/runs/20260902-163914Z-5c68f84-fresh
-created_at: 2026-09-02T16:41:26.327Z
----
-
-# Receipt: collector.request-body-limits — passed
-
-## Observations (expected → seen)
-
-- Exercised 5 production collector POST boundaries: /api/fleet-report/run, /api/fleet-report/import-csv, /api/probe/mta-sts, /api/probe/smtp-starttls, /api/probe/allowlist/generate.
-- 5 declared overflows returned exact 413 {"error":"Request body too large","maxBytes":1048576}.
-- 5 chunked overflows returned exact 413; every stream was cancelled and no sentinel was pulled.
-- 5 UTF-8 byte-overflows had JavaScript length below 1048576 but encoded bytes above it and returned exact 413.
-- 5 bodies were exactly 1048576 bytes and none returned 413; CSV import returned 200 and the four JSON routes kept 400 validation.
-- 5 malformed-under-limit bodies returned 400, preserving route validation.
-
-## Forbidden (must not happen → confirmed absent)
-
-- The enforced outbound guard recorded zero external attempts (providerAttempts count=0; details=[]).
-- No route-specific size error, alternate max value, successful overflow, or post-overflow sentinel read was observed.
-- Declared overflow pull counts were at most one runtime-prefetched chunk; chunked overflow pull counts were at most two.
-
-## Read-back (side effects checked through an independent path)
-
-- Re-read every HTTP exchange from disk and wrote the route/case matrix to `readback/request-body-limit.json`.
-
-## Artifacts
-
-| path | kind | check | sha256 |
-|---|---|---|---|
-| verification/runs/20260902-163914Z-5c68f84-fresh/collector-doctor.txt | txt | aux · unrecognized .txt | 5d5f0b228d14f95a9d842d8f5f09c4bba864011d534db28ce08ddf7d94fa2e67 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/collector-server.log | log | aux · unrecognized .log | 3908c5bb1206094b98681c4ba9935b99c45f19ba45629ba54c09a8b7a4d70d46 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/env.txt | env | aux | ce55727931c10e026ac5a9dec90ab396746591c3d235751b82f18b8915517637 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/healthz.json | json | aux · unrecognized .json | 6b4fcebc70123532d27625947fc9285f88e16d647eda9e66acf99a5160c2c243 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/01-declared-overflow-api-fleet-report-run.json | http | evidence | 006b167866faedbc8eed976aa8ebbf036e7c1f1ab85ee1451f2f3f6f45192e9a |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/02-declared-overflow-api-fleet-report-import-csv.json | http | evidence | 88a4389c8ffd9397c86abde9a73f7b3fa29b6b000f55a06b63e3cceb07a5d743 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/03-declared-overflow-api-probe-mta-sts.json | http | evidence | 8fb12486d623bdf54b18d3ea394b2f671c5efd73a82f688febd0d6f8e6d00e06 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/04-declared-overflow-api-probe-smtp-starttls.json | http | evidence | 74f6b40786b158e95cecb3fd777e14fbd86ab6f0e186563fc68968510d0ff735 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/05-declared-overflow-api-probe-allowlist-generate.json | http | evidence | ec7d57ba69d776fbd3e9547ca8f2323ffed14a5a3928e70751c0226e32e988b4 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/06-chunked-overflow-api-fleet-report-run.json | http | evidence | 4432218f7f639f5fa8644e67cbaef9c09479b91d3b712dcac85fe0df835bf4ed |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/07-chunked-overflow-api-fleet-report-import-csv.json | http | evidence | db908ca4152fc6d54fa4af6359cb22f9870505772a8c5b59e30d51cd51e24c72 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/08-chunked-overflow-api-probe-mta-sts.json | http | evidence | f2d0876f61772117a6a8415982241bd8280510a1ac1e73c455ef924eea90c68a |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/09-chunked-overflow-api-probe-smtp-starttls.json | http | evidence | e3397474826213d4856b073bedc20fa1dadca4806bbdc17d927e371412f8972d |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/10-chunked-overflow-api-probe-allowlist-generate.json | http | evidence | 84182d18e12bbe77ad877ae5f790c7dffd19e88623c1e8100b0462c5f954aecd |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/11-utf8-overflow-api-fleet-report-run.json | http | evidence | af9329d22a73f7ae82e7ddc05f6a43e66cc30fd4161bf95d80e64031b145a64b |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/12-utf8-overflow-api-fleet-report-import-csv.json | http | evidence | 707fd3a6d0d6e067466e442bdb7fa63d1be602d49f4f1205330e99e4a9fc1f30 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/13-utf8-overflow-api-probe-mta-sts.json | http | evidence | a7c20fe259a73a851be5885177d72c4e3555a95217398844e1e88d43989575da |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/14-utf8-overflow-api-probe-smtp-starttls.json | http | evidence | 8483b827f55674f345992128def20f7ba766cbcd57f32188740298009a8b3c58 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/15-utf8-overflow-api-probe-allowlist-generate.json | http | evidence | f1e46e50337532dca14059e5d82a2ddd9f03d12683e99462fc779fd23a7d66d9 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/16-exact-boundary-api-fleet-report-run.json | http | evidence | 459daabb59e06ed7d09240711bbbed59d9d5217ca3e2d66f8ed12ebeda8355a0 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/17-exact-boundary-api-fleet-report-import-csv.json | http | evidence | 8c6f926158d9601cf7bd40b25c4d89b809d2df7ac8fd486217e1425c0b492410 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/18-exact-boundary-api-probe-mta-sts.json | http | evidence | 35b1fedb3c97729f3556a9954a7ed97f88d7689dcde03ac5c71bc8072ecbdf65 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/19-exact-boundary-api-probe-smtp-starttls.json | http | evidence | 68d13e41ef4e67c28ebc320365f483a3af71c002fca94e58c1086f27fd8cd24c |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/20-exact-boundary-api-probe-allowlist-generate.json | http | evidence | 7a162f27557480a5e3517fe497080cf2ff6a793fc82a922a849bf9c3c4c84927 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/21-malformed-under-limit-api-fleet-report-run.json | http | evidence | e54244c453cb6248f15934a070b8185f6fa8db30e38e71176833d8e0f457af79 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/22-malformed-under-limit-api-fleet-report-import-csv.json | http | evidence | baa2877f36964d40047f2b726548e70a65c514a8d585864e62fdf7900cb318f4 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/23-malformed-under-limit-api-probe-mta-sts.json | http | evidence | ae732f439fb5d3f25d0ba1e79a584db3484890262716a7b48a2257b59c7c63df |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/24-malformed-under-limit-api-probe-smtp-starttls.json | http | evidence | e4a91d514c6153cf0556f47257254b714978e70abec6b1e699ab4a70351f0f30 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/http/25-malformed-under-limit-api-probe-allowlist-generate.json | http | evidence | 4175306031590703b06dfb026ef1985b387ba15eb9486b872b875aea150e3720 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/observations.md | md | aux · unrecognized .md | be370242457a16c200769f40b2be4055475712f2ac637815b5cf6da466d213e5 |
-| verification/runs/20260902-163914Z-5c68f84-fresh/readback/request-body-limit.json | readback | evidence | 04e7607b5d1dcbccd8882087ba1ca686d205631cdb287b94af2201440af7510f |
